fix volume modes dep (4,5)
