+++ src/features/Volumes/VolumeDrawer/index.ts
@@ -1,3 +1,4 @@
-import VolumeDrawer, { modes as _modes } from './VolumeDrawer';
+import { modes as _modes } from './modes';
+import VolumeDrawer from './VolumeDrawer';
 export const modes = _modes;
 export default VolumeDrawer;
