+++ src/features/Volumes/VolumeDrawer/modes.ts
@@ -1,0 +1,11 @@
+export const modes = {
+  ATTACHING: 'attaching',
+  CLONING: 'cloning',
+  CLOSED: 'closed',
+  CREATING_FOR_LINODE: 'creating_for_linode',
+  CREATING: 'creating',
+  EDITING: 'editing',
+  RESIZING: 'resizing',
+  VIEW_RESIZE_INSTRUCTIONS: 'viewing_resize_instructions',
+  VIEWING_CONFIG: 'viewing_config'
+};
